refactor: SM-214 Uploader를 base 컴포넌트에서 compound 컴포넌트로 재위치
